Move operator out to its own file
