+++ rule/json.go
@@ -7,80 +7,6 @@ import (
 	"github.com/tidwall/gjson"
 )
 
-type operator struct {
-	kind     string
-	operands []Expr
-}
-
-func (o *operator) PushExpr(e Expr) {
-	o.operands = append(o.operands, e)
-}
-
-func (o *operator) Same(c ComparableExpression) bool {
-	if o.kind == c.GetKind() {
-		o2, ok := c.(operander)
-		if ok {
-			ops := o2.Operands()
-			len1 := len(o.operands)
-			len2 := len(ops)
-			if len1 != len2 {
-				return false
-			}
-			for i := 0; i < len1; i++ {
-				ce1 := o.operands[i].(ComparableExpression)
-				ce2 := ops[i].(ComparableExpression)
-				if !ce1.Same(ce2) {
-					return false
-				}
-			}
-			return true
-		}
-		return false
-	}
-	return false
-}
-
-//
-func (o *operator) GetKind() string {
-	return o.kind
-}
-
-func (o *operator) UnmarshalJSON(data []byte) error {
-	var node struct {
-		Kind     string
-		Operands operands
-	}
-
-	err := json.Unmarshal(data, &node)
-	if err != nil {
-		return err
-	}
-
-	o.operands = node.Operands.Exprs
-	o.kind = node.Kind
-
-	return nil
-}
-
-func (o *operator) MarshalJSON() ([]byte, error) {
-	var op struct {
-		Kind     string `json:"kind"`
-		Operands []Expr `json:"operands"`
-	}
-
-	op.Kind = o.kind
-	op.Operands = o.operands
-	return json.Marshal(&op)
-}
-
-func (o *operator) Eval(params Params) (*Value, error) {
-	return nil, nil
-}
-
-func (o *operator) Operands() []Expr {
-	return o.operands
-}
-
 type operands struct {
 	Ops   []json.RawMessage `json:"operands"`
 	Exprs []Expr
